qmljs/parser: update to latest qt5 parser
fix error semicolon insertion of loops, strict mode

+++ src/libs/qmljs/parser/qmldirparser_p.h
@@ -128,7 +128,7 @@ public:
 #endif
 
 private:
-    void reportError(int line, int column, const QString &message);
+    void reportError(quint16 line, quint16 column, const QString &message);
 
 private:
     QList<QmlError> _errors;
